+++ driver-foundationdb/src/main/java/io/sbk/FoundationDB/FoundationDBMultiKeyReader.java
@@ -70,7 +70,7 @@ public class FoundationDBMultiKeyReader implements Reader<byte[]> {
             throws EOFException, IOException {
         final int recs;
         if (params.getRecordsPerReader() > 0 && params.getRecordsPerReader() > cnt) {
-            recs = Math.min(params.getRecordsPerReader() - cnt, params.getRecordsPerSync());
+            recs = (int) Math.min(params.getRecordsPerReader() - cnt, params.getRecordsPerSync());
         } else {
             recs =  params.getRecordsPerSync();
         }
@@ -105,7 +105,7 @@ public class FoundationDBMultiKeyReader implements Reader<byte[]> {
             throws EOFException, IOException {
         final int recs;
         if (params.getRecordsPerReader() > 0 && params.getRecordsPerReader() > cnt) {
-            recs = Math.min(params.getRecordsPerReader() - cnt, params.getRecordsPerSync());
+            recs = (int) Math.min(params.getRecordsPerReader() - cnt, params.getRecordsPerSync());
         } else {
             recs =  params.getRecordsPerSync();
         }
