+++ driver-ignite/src/main/java/io/sbk/Ignite/IgniteTransactionWriter.java
@@ -30,7 +30,7 @@ public class IgniteTransactionWriter implements Writer<byte[]> {
     private final IgniteCache<Long, byte[]> cache;
     private final org.apache.ignite.Ignite ignite;
     private long key;
-    private int cnt;
+    private long cnt;
 
     public IgniteTransactionWriter(int id, Parameters params, IgniteCache<Long, byte[]> cache,
                                    org.apache.ignite.Ignite ignite) throws IOException {
@@ -59,7 +59,7 @@ public class IgniteTransactionWriter implements Writer<byte[]> {
     public void writeAsyncTime(DataType<byte[]> dType, byte[] data, int size, Time time, Status status) throws IOException {
         final int recs;
         if (params.getRecordsPerWriter() > 0 && params.getRecordsPerWriter() > cnt) {
-            recs = Math.min(params.getRecordsPerWriter() - cnt, params.getRecordsPerSync());
+            recs = (int) Math.min(params.getRecordsPerWriter() - cnt, params.getRecordsPerSync());
         } else {
             recs = params.getRecordsPerSync();
         }
@@ -82,7 +82,7 @@ public class IgniteTransactionWriter implements Writer<byte[]> {
                             Status status, SendChannel sendChannel, int id) throws IOException {
         final int recs;
         if (params.getRecordsPerWriter() > 0 && params.getRecordsPerWriter() > cnt) {
-            recs = Math.min(params.getRecordsPerWriter() - cnt, params.getRecordsPerSync());
+            recs = (int) Math.min(params.getRecordsPerWriter() - cnt, params.getRecordsPerSync());
         } else {
             recs = params.getRecordsPerSync();
         }
