+++ driver-mongodb/src/main/java/io/sbk/MongoDB/MongoDBMultiWriter.java
@@ -29,7 +29,7 @@ public class MongoDBMultiWriter implements Writer<byte[]> {
     final private MongoCollection<Document> databaseCollection;
     final private Parameters params;
     private long key;
-    private int cnt;
+    private long cnt;
 
     public MongoDBMultiWriter(int id, Parameters params, MongoDBConfig config,
                               MongoCollection<Document> databaseCollection) throws IOException {
@@ -60,7 +60,7 @@ public class MongoDBMultiWriter implements Writer<byte[]> {
     public void writeAsyncTime(DataType<byte[]> dType, byte[] data, int size, Time time, Status status) throws IOException {
         final int recs;
         if (params.getRecordsPerWriter() > 0 && params.getRecordsPerWriter() > cnt) {
-            recs = Math.min(params.getRecordsPerWriter() - cnt, params.getRecordsPerSync());
+            recs = (int) Math.min(params.getRecordsPerWriter() - cnt, params.getRecordsPerSync());
         } else {
             recs = params.getRecordsPerSync();
         }
@@ -84,7 +84,7 @@ public class MongoDBMultiWriter implements Writer<byte[]> {
                             Status status, SendChannel sendChannel, int id) throws IOException {
         final int recs;
         if (params.getRecordsPerWriter() > 0 && params.getRecordsPerWriter() > cnt) {
-            recs = Math.min(params.getRecordsPerWriter() - cnt, params.getRecordsPerSync());
+            recs = (int) Math.min(params.getRecordsPerWriter() - cnt, params.getRecordsPerSync());
         } else {
             recs = params.getRecordsPerSync();
         }
